+++ src/main/java/com/it/every/student/model/StudentService.java
@@ -24,4 +24,10 @@ public interface StudentService {
 	
 	List<Map<String, Object>> selectByStuId (String userid);
 	Map<String, Object> selectByTuitionNo (String userid ,int tuitionNo);
+	
+	
+	List<Map<String, Object>> selectByScholarship (String userid);
+	
+	
+	
 }
