+++ src/main/java/config/mybatis/mapper/oracle/student.xml
@@ -44,10 +44,14 @@
 	
 	<!-- 장학금 조회 -->
 	<select id="selectByScholarship" parameterType="String" resultType="map">
-		select * from scholarshipView
-		where student_id=#{123} 
+			select *
+			from scholarshipView a
+			join student s
+			on a.stu_no = s.stu_no
+			where s.student_id=#{123}
 	</select>
-
+	
+	
 	
 </mapper>
